util: avoid image_error() segfault if no handler
If an image has no handler yet and image_error() is called, genimage
segfaults on image->handler->type. This case happens for example in
image_set_handler() if the config didn't provide any type for an image.

To fix this, first check the image handler, or use the type "unknown".

Signed-off-by: Vivien Didelot <vivien.didelot@savoirfairelinux.com>
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>