image-hd: set the position of extended partition
The hdimage property "extended-partition" sets the position of the
extended partition recorded into the Master Boot Record.

Setting the extended partition to 2, limits the number of primary
partitions to 1 into the MBR. It allows to create Extended Partition
Boot Record scheme images. This kind of images is composed of one
primary and one extended partition. Extra partitions are located into
the extended partition.

Signed-off-by: Gaël PORTAY <gael.portay@savoirfairelinux.com>