systemp: Print command output on demand
The output of invoked commands can be very useful for debugging, but
also annoying or confusing. Print stderr of invoked commands on
debug loglevel and and stdout of invoked commands on verbose debug
loglevel, but redirect to /dev/null on lower loglevels. This
makes the manual > /dev/null redirection in some commands unnecessary
which is removed with this patch.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>
[mol: fix error handling, change output loglevel, log stdout in the tests]
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>