genimage: Use absolute pathes
Relative pathes can be error prone when it's not clear to
which path the pathes are relative to. For FIT image support
we have to create pathes that are not relative to the current
directory, but to the path where the current its input file is.

This patch lets imagepath(), inputpath(), rootpath() and tmppath()
return absolute filenames rather than directly the (absolute or
relative) input pathes.

Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>