test: use partitions with actual content
The partitions where accidentally created with /dev/null so the size was
zero. Use /dev/zero to get the expected size. Update all test results to
match this.
Rewrite the qemu check. Different versions of qemu-img produce different
image sizes so let qemu-img compare the image to a reference file.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>