image-hd: add support for bootable flags on GPT
Bit 2 of the GPT partition table entry "flags" attribute allows to
mark the partition as bootable, so we add support for setting this
flag, using the same bootable flag used for MBR partition tables.

Setting this flag is useful to properly support the "Generic Distro"
concept from U-Boot [1], which searches for boot scripts in the
bootable partition.

[1] http://git.denx.de/?p=u-boot.git;a=blob;f=doc/README.distro

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>