hdimage: Implement 'gpt-location' option to move GPT table
This allows the user to specify the location of the GPT table.  This is
useful if a device requires that the bootloader be located in a location
that conflicts with the normal location of the GPT table.

Also, this changes the calculation of the "first usable LBA" to be the
first *listed* partition; this allows the calculation to disregard the
non-listed bootloaders that make this feature useful.  It is up to the
user to ensure that the GPT table is in a location that does not
conflict with any non-listed partition entries.

Signed-off-by: George Hilliard <ghilliard@kopismobile.com>