image-ext: Pass 'features' to mke2fs
genimage does not pass the config options 'features' to mke2fs when
option 'use-mke2fs' is set to true. If the latter option is not set,
genext2fs is called instead and tune2fs is run on the image afterwards
passing those features.

By not passing that features the behaviour contradicts the documentation
and the user expectations.

Fixes: #81
Signed-off-by: Alexander Dahl <ada@thorsis.com>