util: pad_file: always use the full image size for the size argument
Before this, the interpretation of the 'size' argument depended on the
'infile'. If no 'infile' was given, then 'size' was the final size of the
total image.
If an 'infile' was given, then 'size' was the number of bytes to add to the
images.

This is confusing, so change it to always mean the full size the image is
supposed to be once pad_file() is done.

Fixes: adf18f077010 ("image-flash: Pad empty flash partitions to size")
Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>