image-rauc: allow optional keyring verification
A keyring can be specified in genimage's config similar to "cert" and
"key". The keyring is passed to the "rauc bundle" command which makes
sure the signed bundle can be verified against that keyring.

This is especially helpful for use cases that exceed a single key pair
with self-signed certificate, such as a PKI or a keyring consisting of
multiple certificates.

Signed-off-by: Bastian Krause <bst@pengutronix.de>