map_file_extents: explicitely use output of ioctl
Even though the file and its extend mapping should not be modified between both
ioctl()'s there is the theoretical possibility that it was and less extends than
expected are returned. Hence, use the ioctl()'s output and not its input
parameter further on (see [1]), although they should have the same value.

[1] https://elixir.bootlin.com/linux/v5.7/source/Documentation/filesystems/fiemap.txt#L62

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>