genimage: make sure that IMAGEMOUNTPATH is exported correctly
If the image mountpoint is not explicitly defined, then IMAGEMOUNTPATH is
currently no exported. This is incorrect, because the default mountpoint
for any image is '/'.
Use mountpath() to get the correct path. It handles the default mountpoint
correctly.
Make sure that IMAGEMOUNTPATH is not set if the image is explicitly marked
as empty.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>