allow specifying an image as temporary
When using something like

image @IMAGE@ {
    ....
    partition sysdata {
        image = "sysdata"
	in-partition-table = true
	partition-type-uuid = "L"
    }
}

image sysdata {
    size = 200M
    ext4 {}
    empty = true
}

to create a partition containing an empty file system, it's
undesirable to leave such an empty file system image in the output
directory (especially when building with Yocto, where it then gets
copied to the shared deploy directory).

So allow specifying an image as temporary, putting it in in tmppath()
rather than imagepath().

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>