add macro for compile-time assertions
Compile-time assertions can be helpful to ensure the compiler actually
lays out sturcts as one expects, and for a lot of other things as
well.

<assert.h> exposes a static_assert() macro that maps to _Static_assert
if one compiles with -std=c11 or later, but I don't think we can
mandate that. Also, that standard macro suffers from the defect of
requiring a separate string message - it's usually completely obvious
from the stringification of the expression what it is that is being
asserted. So just define our own ct_assert().

[Side note: We seem to be relying implicitly on gcc defaulting to some
gnuXX mode - building with -std=c99 or -std=c89 breaks with complaints
about the use of typeof().]

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>