image-hd: allow individual alignment setting for each partition
When building an image that has a combination of some oddball smallish
partitions (say, for containing a U-Boot environment, some firmware
blob, or anything else that just needs a few KB), and a bunch of
partitions meant to hold normal file systems, it can be useful to
ensure those filesystems get aligned on a MB boundary, while not
forcing all the small partitions to effectively occupy a MB each - in
some cases that's not even possible, if for example the artifacts they
contain must all be found within the first MB of the disk.

So allow each partition individually to specify the alignment it
should have. If offset is given explicitly, align is ignored.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>