image-hd.c: don't lazily initialize now
The code is easier to read and reason about if the "now" variable is
initialized before the loop.

Note that with the previously added sanity check, the condition
part->in_partition_table implies hd->partition_table, so the if() that
gets deleted here could just have been spelled "if (!now)".

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>