image-hd: expand the image to the full size after writing the partition table
If 'fill' is enabled, the the image is expanded to the full image size.
When the GPT partition table is written, genimage tries to append it at the
end of the file (with padding if necesssary). This fails if the image is
expanded first.
So write the partition tables first and then expand the image as necessary.
This fixes using 'gpt = true' and 'fill = true' for an image.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>