image-hd.c: do not require the autoresize partition to be specified last
There's no requirement that partitions are listed in order in the
partition table, and one may in fact (for legacy reasons) want certain
partitions to have particular numbers. So drop the requirement that an
autoresize partition is the last one - of course, any partitions
specified later in the configuration will need to have an explicit
offset set in order not to overlap with the autoresize partition.

This can be useful to e.g. give a name (and make them appear in the
partition table) to areas of the disk that were previously only
accessed via the raw device (say, bootloader or U-Boot environment),
without altering the enumeration of existing partitions.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>