image-hd: stricter option handling
There have been situations where the hybrid mode was accidentally used
because a config contained a bogus partition-type entry.

So lets make this explicit: There is now a new 'partition-table-type'
option where the hybrid mode must be selected explicitly.
The relevant options for mbr and gpt are only acceptable if the
corresponding type (or hybrid) is selected.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>