image-hd.c: separate computation of part->offset from the sanity checks
The code is easier to read and reason about if one clearly separates
the determination of the part->offset value from the sanity checks
done to it.

Yes, in many cases the way part->offset is computed will make those
sanity checks automatically fulfilled, but this in no way
performance-critical code, so let's just do them always.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>