image-hd.c: always sanity check child->size v part->size
Yes, if we entered the preceding if(), the resulting part->size is
obviously big enough for child->size.

But clearly separating the automagic computations of default values
from the sanity checks makes the code easier to read (in particular,
makes it easier to see that the invariants we claim are actually
enforced).

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>