image-hd.c: move overlap check below partition size determination
We'll later change the overlap check to allow placing partitions
before "now" (but of course still not overlapping existing ones). So
move the check to after both the offset and the size of the partition
are known. No functional change, although an erroneous image
definition may now fail with one error instead of another.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>