image-hd.c: allow partitions to appear out-of-order
GPT partitions, and primary partitions when using MBR, do not need to
appear in order. This can be useful to add new partitions to the
partition table (e.g. including an explicit bootloader partition,
instead of relying on writing it to a raw offset on the block device),
without altering the numbers of previously existing partitions.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>