image-hd.c: fix missing check for overlap with GPT table
I consider this a somewhat elegant solution: By making sure there are
"partitions" in the list representing the areas occupied by the
MBR (the tail part of it at least), GPT header and GPT array, all the
overlap checking automatically Just Works.

When the image size is known up front, we also know where the backup
GPT header/array will be placed and can add a corresponding fake
partition.

With this, we now ensure that no partitions get placed on top of the
GPT array or other partition table metadata.

These are all added at the end of the list of partitions: We will
later add support for images having known "holes" or "don't care"
regions, where it is ok for this image metadata to be placed. For that
to work, we must do the overlap checking in the same order as the
contents is going to be written to the final image.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>