image-file.c: allow specifying "don't care" regions
Some bootloader images are specially crafted to avoid having any
meaningful code/data in the range from where the MBR table starts (at
offset 440) and to beyond where the GPT array ends (33K), or perhaps
just (if only meant to be used with MBR disks) a hole at [440,512].

In order to use such images, while still doing overlap sanity checks,
one needs a way to tell genimage that certain regions of a file are
not significant and can thus be allowed to be overwritten.

I decided to make this a property of the image rather than something
to specify inside the partition that is to contain the image, since
this is fundamentally a property of the image/the file in question.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>