test: add tests for "bootloader with holes allowing some overlap"
The image_setup logic is alright, but we still fail the last test case
during image generation. The problem is that after the bootloader
partition has been emitted to the output file (at offset 128K), the
output file is then 198K large. And when pad_file() is then called a
little later to add the 3K partition at offset 129K, pad_file() is not
able to cope with the output already being big enough; we fail due to
the

	if ((unsigned long long)s.st_size > size) {
		image_error(image, "output file '%s' too large\n", outfile);
		ret = -EINVAL;
		goto err_out;
	}

in the first call of pad_file() from hdimage_generate().

But just changing that logic to a "hm, ok, no need to do any padding"
isn't enough, because there's also a lseek(SEEK_END) v
image->last_offset sanity check, and the calling convention for
pad_file simply doesn't allow writing the image to anything but
image->last_offset (i.e., at the current end).

This all requires a bit more surgery - it's probably simplest to
create a new insert_file() function similar to insert_data() which is
capable of adding data anywhere in the output image, and not rely on
always adding new content at the end. For now, let's just have the
known-broken test case remind us to fix this.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>