image-hd.c: prepare for fake partitions
We're going to add a few "fake" partitions at the end of the list of
partitions in order to represent areas of the disk that must overlap
any real partitions. Make sure those are not confused with an extended
partition, and that the logic in hdimage_insert_ebr which looks at the
next partition (if any) to create the second EBR entry ignores any
partition which is not an extended partition.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>