util.c: add insert_image() helper
Add a helper that can be used to insert one image into another at a
specific offset. The caller must also pass an explicit size; if the
inserted image is smaller than this (including passing NULL for no
image), the remainder is filled with zero bytes.

This is to transition image-hd.c away from pad_file(), which has a
somewhat awkward interface and assumes the output image is always
built up sequentially.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>