image-hd.c: truncate output file before looping over partitions
The logic currently relies on the first call of pad_file() being
passed MODE_OVERWRITE which in turn causes the open() to be passed
O_TRUNC, while subsequent calls of pad_file() then get MODE_APPEND.

We will eliminate use of pad_file completely, and the new helper
insert_image() always just opens the file with O_CREAT|O_WRONLY, so we
need to ensure that the output file, if it exists, has no contents.

Presumably unlink() could be used with mostly the same effect - except
if the output filename is a (possibly dangling) symlink, so I decided
to use truncate().

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>