image-hd.c: correct computation of first_usable_lba
Since we now allow partitions to appear in any order, the
first_usable_lba should not be computed as the start of the first
listed partition in the partition table. Instead, use the start of
whatever partition in the table that has the smallest offset.

To simplify the if() condition a little I moved the "ignore this
partition if it's not listed" check up.

I considered instead just unconditionally using the LBA immediately
following the primary GPT array. However, if there is a bootloader
spanning the MBR+GPT header+GPT array, the area on the other side of
the GPT array should not be available for adding a new partition. And
doing it this way is in any case more aligned with what was done
previously.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>