image-hd.c: ignore in-partition-table when the image has no partition table
If somebody sets "partition-table = false", we might as well ignore
the "in-partition-table" setting and force it to false for all
partitions (it is true by default).

I previously tried adding a sanity check for the combination
"partition-table = false", "in-partition-table = true" and rejecting
that, but it was rightly pointed out that that isn't very ergonomic.
If somebody uses image-hd to just create an image with certain blobs
at certain offsets, they should not have to meticulously add
"in-partition-table = false" for each partition - the setting in the
hdimage{} section clearly states their intent.

If libconfuse at some point gives us a way to ask whether a value is
present or we were handed back the default value, one could add such a
sanity check for the case of somebody explicitly setting
"in-partition-table = true".

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>