image-hd.c: make GPT backup handling a little less special
The handling of the GPT backup header/array is by its nature a little
special.

- When the image size is known, we know upfront where it is going to
be.

- If the image size is determined by the contents, the backup should
almost, but not quite, follow the normal rule of "put it at 'now'
rounded up to a suitable alignment". The problem is that it is the
_end_ of this fake partition that should be suitably aligned.

Still, we can simplify things a little by always having the fake
partition present (instead of conditional on image->size != 0), with
the normal convention of offset==0 meaning "determined dynamically",
then just teach the code which handles offset==0 to treat the
gpt_backup partition specially.

That in turn means that, at the place where we determine the final
image size, we no longer need to distinguish hd->gpt or not; the
placement of the backup gpt has caused 'now' to have precisely the
right value.

That gpt_backup now exists whenever hd->gpt is true also means we can
replace the code in the autoresize logic which reserves room for the
backup to actually refer to gpt_backup, which makes it a little more
self-explanatory than the magic subtraction of GPT_SECTORS*512.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>