image-hd.c: fix extended partition with offset
If the first extended partition had an offset configured,
the offset to the second extended partition would not be calculated
correctly.
Also, the size calculation for the extended partition is not quite correct.

Fix all this and make it more explicit with the following assumtions:

- There are no primary partition after the extended partion, so the image
  size can be used to calculate the size of the extended partition.
- The EBR before each logical partition is always 'hd->align' bytes before
  the start of this partition.

Based on a PR from Ruben Vandamme <github-com-00ff86@vandamme.email>.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>