image-hd: resize the image to the final size at the begining
This is not strictly necessary but this circumvents XFS preallocation
heuristics. Without this, the holes in the image may be smaller
than necessary.
Check that the image still has the same size. This is just a sanity check
to validate that the image did not grow past the initially calculated size.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>