insert_image: fix read/write loop
Update the offset with the actual written size. It may be shorter, if the
read returned less than 'now'.
And abort when the end of the file is reached. This is not yet a problem,
because in hdimage the actual file size is used as size. But other
use-cases will be different.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>