add --includepath option
When ${B} was made distinct from ${S} and ${WORKDIR} in
genimage.bbclass in meta-ptx commit 9eaa29b7, use of include() in the
main configuration file to include other fragments from SRC_URI broke.

It can be worked around in a recipe by copying all the fragments to
${B} in a do_genimage[prefunc], but that's awkward. It's more natural
for genimage to take a --includepath option and genimage.bbclass to
pass --includepath ${S}.

This does require libconfuse >= 2.8, while meta-ptx currently has a
recipe for 2.7 (the last release downloadable from savannah, but the
project has moved to github).

confuse.h says:

 * Note that the current directory is not added to the searchpath
 * by default.

That's probably not expected, so when this option is used, always add
"." as well.

In order for this to be meaningful in a config{} section inside the
configuration file, one would have to get libconfuse to call a custom
function when encountering this; it cannot just be a simple
CFG_STR. Because if it doesn't take effect immediately, parsing the
rest of the config file will most likely fail when it hits the
include()s. But the whole option parsing machinery is basically built
around the assumption that options take single string arguments (which
is also why this had to accept a colon-separated list instead of
supporting --includepath being given multiple times). I don't think
it's a huge loss, C files also don't have the proper values of -I
embedded into them. But it does require relaxing the wording in
README.rst a bit.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>