config.c: allow hiding specific command line options from --help
Since --includepath will be conditionally supported, add a flag to
struct config so we can suppress it from --help if genimage was built
against an old libconfuse. We will still accept the option and then
display an error message better than "invalid option --includepath",
but there's no reason to advertise it in --help.

This also makes writing a test case a little easier - or rather, the
condition for when to skip the test.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>