image-squashfs: handle image->size appropriately
I was trying to actually make use of genimage's ability to split a
root filesystem into a read-only part stored in a squashfs, and a
writable /home partition in an ext4 image. Unfortunately, the squashfs
didn't end up included in the sd image:

INFO: hdimage(...): adding partition 'rootfs-A' (in MBR) from 'rootfs.squashfs' ...
INFO: hdimage(...): adding partition 'rootfs-B' (in MBR) from 'rootfs.squashfs' ...

INFO: hdimage(...): adding partition 'home' (in MBR) from 'home.ext4' ...
DEBUG: hdimage(...): copying 524288000 bytes from .../genimage-tmp/home.ext4 at offset 1577058304

I.e., there's no "copying xxx bytes from" lines following the "adding
partition rootfs" lines. That's due to

		if (child->size == 0)
			continue;

around line 457 in image-hd.c.

For a squashfs, we don't know the final size of the image, and we
should not be forced to decide upfront how large it can be. So
interpret a squashfs image node's size, when given, as a maximum
size. So if a size has been set and the generated image turns out to
be larger, make it an error.

Otherwise, that is, when the size is given and >= the generated
image, or when the size is not given explicitly, set image->size to
the size of the generated file.

There is one small problem: Consumers, e.g. image-hd.c, may have done
a sanity check (part size v child image size) during the setup phase,
but we don't (and can't) set the final size until the generate
phase. The next patch will fix that by repeating the sanity check
during generate.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>