collect_mountpoints(): only do timestamp fixup when needed
When trying to use genimage on an embedded target, I hit

  find: unrecognized: -printf
  xargs: invalid option -- '0'

because the busybox I have apparently has somewhat limited "find" and
"xargs" applets. OTOH, I'm not using the "split a rootfs"
functionality - in fact, I'm just pointing rootpath at an empty dir
which won't be used for anything, so the whole thing is redundant.

When there are no non-trivial mountpoint=... options, we might as well
skip the timestamp fixup.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>