image-hd.c: use size of destination block device as image size
Currently, when writing a GPT image directly to a block device, the
image generation fails when hdimage_insert_gpt() calls extend_file(),
because the computed size of the image (based on the defined
partitions etc.) is smaller than what extend_file() sees when it does
the SEEK_END sanity check.

So when the target is a block device and the image size has not been
set explicitly, set image->size from the actual size of the block
device. In case one has a few board variants with slightly differing
eMMC sizes, that will automatically DTRT - we still do all the sanity
checks that the defined partitions do not exceed the image size etc.

In fact, when writing directly to a block device, we are implicitly
populating the whole device - and in particular, when writing a GPT
image, the backup header must be placed at the end of the
device. There is currently no known use case for setting a non-zero
image size (and having genimage check that is <= the actual size
etc...), so for now enforce that image->size is not set for a block
device target.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>