test: mke2fs: fully disable quota in the test
The test disabled quota for better reproducibility. But it seems some quota
stuff remains. And at leasts with mke2fs 1.46.2 on x86_64 the resulting
image seems to be inconsistent. At least e2fsck complains about a quota
issue:

[QUOTA WARNING] Usage inconsistent for ID 0:actual (29696, 42) != expected (13312, 2)
mke2fs: Update quota info for quota type 0.
[QUOTA WARNING] Usage inconsistent for ID 0:actual (29696, 42) != expected (13312, 2)
mke2fs: Update quota info for quota type 1.

For some reason this does not happen on 32 bit architectures. So the
resulting image is different and the check fails.

Avoid this whole mess by adding '-E quotatype=' to the mke2fs arguments.
This way quota is fully disabled, e2fsck is happy and the resulting
filesystem is the same on all architectures.

Fixes: #200

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>