image-hd: implement "fill" option for partitions
When bootstrapping, it can be useful to ensure that a partition has
completely known contents, especially when the partition just contains
raw binary data and not a filesystem.

When generating an image file, this is implicitly done already because
the non-written parts of the partition(s) are implicitly zero-filled
when the containing image file is extended to its final size.

But when using genimage to directly populate a block device, no such
implicit zero-filling is done.

Add a "fill" boolean option, and use that to decide between passing
the partition's size or the child image's size to insert_image().

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>