README: Formalize signing off commits
The automatic checks for pull requests on Github include a check that
fails if a Signed-off-by: line by the commiter is missing in any commit.

It's supposed to mean the "usual semantic also used in other open-source
projects" but so far that is not spelled out anywhere.

So add a section to the README to make this explicit.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>