image-hd: split out partition image handling and fix the file_size
The last partition may not be the one at the end of the image, so it
could happen, that file_size is overwritten with a smaller value.

It's usually not notiable because the value is only used for the initial
image size. The file can grow larger when it is actually filled.

Signed-off-by: Michael Olbrich <m.olbrich@pengutronix.de>